make homepage fragment instead of activity
