+++ app/src/main/res/layout/homepage_fragment.xml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.example.tolearn.ProfileFragment"
+    >
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+        <TextView
+            android:id="@+id/joinedClassesTV"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Joined Classes"
+            android:gravity="center"
+            android:background="@drawable/border_shadow_white_background"
+            android:textSize="22sp"
+            android:textColor="@color/blue_learning"
+            ></TextView>
+        <ScrollView
+            android:id="@+id/joinedClassesScroll"
+            android:layout_width="match_parent"
+            android:layout_height="280dp"></ScrollView>
+        <TextView
+            android:id="@+id/createdClassesTV"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Created Classes"
+            android:gravity="center"
+            android:background="@drawable/border_shadow_white_background"
+            android:elevation="15dp"
+            android:textSize="22sp"
+            android:textColor="@color/blue_learning"
+            android:layout_marginBottom="10dp"
+            ></TextView>
+
+        <com.facebook.shimmer.ShimmerFrameLayout
+            android:id="@+id/shimmerLayout"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            tools:ignore="MissingConstraints">
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <!--add  several shimmer placeholder layout -->
+
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+                <include layout="@layout/shimmer_placeholder_layout"></include>
+
+            </LinearLayout>
+        </com.facebook.shimmer.ShimmerFrameLayout>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/createdClassList"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="horizontal"
+            android:divider="@android:color/transparent"
+            android:dividerHeight="5dp"></androidx.recyclerview.widget.RecyclerView>
+
+    </LinearLayout>
+
+
+</RelativeLayout>
