+++ app/src/main/res/menu/drawer_menu.xml
@@ -4,9 +4,9 @@
     tools:showIn="navigation_view">
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_message"
-            android:icon="@drawable/ic_message"
-            android:title="Message" />
+            android:id="@+id/homepage"
+            android:icon="@drawable/ic_baseline_home_24"
+            android:title="homepage" />
         <item
             android:id="@+id/nav_chat"
             android:icon="@drawable/ic_chat"
